requirements.txt
streamlit
PyPDF2
transformers
torch
sentencepiece
requests
scikit-learn
sentence-transformers
matplotlib
python-dotenv
mcp
